cache/requirements.txt
flask
redis
pillow
requests
tornado